create useCountdown hook
